third_party/blink/web_tests/platform/generic/fast/forms/button/button-in-forms-collection-expected.txt

PASS document.test.elements.length is 2
PASS document.test.elements[0].name is "foo"
PASS document.test.elements[1].name is "bla"
PASS document.test.foo.value is "Button1"
PASS document.test.bla.value is "Button2"
PASS successfullyParsed is true

TEST COMPLETE


The following tests for regression against 4405: <button> not listed in form collections. You should see a series of lines starting with PASS.

 Button2
